requirements.txt
numpy
spacy
pyvi
matplotlib
deep-translator
datasets
contractions
torchmetrics
nltk